requirements.txt
anthropic
python-dotenv
video-to-mp3
SpeechRecognition
pydub
requests
openai
flask